Send orderly container an interrupt before stopping
As documented in VIMC-805, the orderly container will respond quickly
to an interrupt, but not to whatever docker stop (via docker-compose stop)
is sending.  This is (presumably) a limitation of httpuv and not something
I can see how to work around at the R level.  So instead we send an
interrupt signal (SIGINT) just before the stop, and that seems to bring
things down much more quicky.