Locate submodules when called out of montagu tree
Currently the deploy code all assumes that we're running from the montagu
directory; this works for most of what we do but for the science restore
we need to run from wherever cron wants.

We already do some working directory funny business in the main montagu
script; this problem differs though because it needs to work at the point that
modules are imported, not when the code is run.